chore: update bundle size numbers for v3-preview
