+++ README.md
@@ -19,14 +19,14 @@
 - @aws-sdk/v3-preview (VERSION: @aws-sdk/client-dynamodb-node@0.1.0-preview.9)
 
   ```console
-          Asset     Size  Chunks             Chunk Names
-  createNote.js  154 KiB       0  [emitted]  createNote
-  deleteNote.js  154 KiB       1  [emitted]  deleteNote
-     getNote.js  154 KiB       2  [emitted]  getNote
-  updateNote.js  154 KiB       3  [emitted]  updateNote
+  $ ls -lh dist | awk '{print $6,$10}'
+  153K createNote.js
+  153K deleteNote.js
+  153K getNote.js
+  153K updateNote.js
   ```
 
-- @aws-sdk/v3-beta (VERSION: @aws-sdk/client-dynamodb@1.0.0-beta.1)
+- @aws-sdk/v3-beta (VERSION: @aws-sdk/client-dynamodb@1.0.0-beta.3)
 
   ```console
   $ ls -lh dist | awk '{print $6,$10}'
