+++ packages/v3-beta/src/getClient.js
@@ -1,3 +1,3 @@
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 
-export const getClient = () => new DynamoDBClient();
+export const getClient = () => new DynamoDBClient({});
