+++ packages/v3-preview/package.json
@@ -1,9 +1,10 @@
 {
   "name": "@aws-sdk/v3-preview",
   "version": "0.1.0",
-  "main": "index.js",
+  "main": "./src/index.js",
   "license": "MIT",
   "dependencies": {
     "@aws-sdk/client-dynamodb-node": "^0.1.0-preview.9"
-  }
+  },
+  "sideEffects": false
 }
