+++ packages/v3-preview/src/deleteItem.js
@@ -1,0 +1,5 @@
+import { getClient } from "./getClient";
+import { DeleteItemCommand } from "@aws-sdk/client-dynamodb-node";
+
+export const deleteItem = async params =>
+  getClient().send(new DeleteItemCommand(params));
