+++ packages/v3-preview/src/getClient.js
@@ -1,0 +1,3 @@
+import { DynamoDBClient } from "@aws-sdk/client-dynamodb-node";
+
+export const getClient = () => new DynamoDBClient({});
