+++ packages/v3-preview/src/getItem.js
@@ -1,0 +1,5 @@
+import { getClient } from "./getClient";
+import { GetItemCommand } from "@aws-sdk/client-dynamodb-node";
+
+export const getItem = async params =>
+  getClient().send(new GetItemCommand(params));
