+++ packages/v3-preview/src/putItem.js
@@ -1,0 +1,5 @@
+import { getClient } from "./getClient";
+import { PutItemCommand } from "@aws-sdk/client-dynamodb-node";
+
+export const putItem = async params =>
+  getClient().send(new PutItemCommand(params));
