+++ packages/v3-preview/src/updateItem.js
@@ -1,0 +1,5 @@
+import { getClient } from "./getClient";
+import { UpdateItemCommand } from "@aws-sdk/client-dynamodb-node";
+
+export const updateItem = async params =>
+  getClient().send(new UpdateItemCommand(params));
